Temperature as 16 bit

+++ Inc/main.h
@@ -141,7 +141,7 @@ typedef struct
 	uint8_t 		char_dyn_adc_state;
 	uint8_t 		assist_level;
 	uint8_t 		regen_level;
-	int8_t         Temperature;
+	int16_t         Temperature;
 	int8_t         	system_state;
 	int8_t         	gear_state;
 	int8_t         	error_state;
